+++ ChangeLog.md
@@ -6,6 +6,8 @@
   - displayLineSegments: fix variable name and remove unused variables.
    (Phuc Ngo  [#59](https://github.com/DGtal-team/DGtalTools-contrib/pull/59))
 
+- *Geometry3d*
+   - volFillCCSize: new tool to fill each Connected Component with a value corresponding to the number of voxels of the CC.   
 
 
 # DGtalTools-contrib  1.2